+++ src/main/java/ma/vi/esql/syntax/expression/NegatableDoubleSubExpressions.java
@@ -5,11 +5,16 @@
 package ma.vi.esql.syntax.expression;
 
 import ma.vi.base.tuple.T2;
+import ma.vi.esql.database.EsqlConnection;
+import ma.vi.esql.exec.env.Environment;
 import ma.vi.esql.semantic.type.Type;
 import ma.vi.esql.semantic.type.Types;
 import ma.vi.esql.syntax.Context;
 import ma.vi.esql.syntax.Esql;
 import ma.vi.esql.syntax.EsqlPath;
+import org.pcollections.PMap;
+
+import static ma.vi.esql.translation.SqlServerTranslator.requireIif;
 
 /**
  * Abstract parent of ESQL expressions taking exactly two arguments.
@@ -51,6 +56,18 @@ public abstract class NegatableDoubleSubExpressions<V> extends Expression<V, Str
   @Override
   public Type computeType(EsqlPath path) {
     return Types.BoolType;
+  }
+
+  @Override
+  protected String trans(Target               target,
+                         EsqlConnection esqlCon,
+                         EsqlPath             path,
+                         PMap<String, Object> parameters,
+                         Environment env) {
+    boolean sqlServerBool = target == Target.SQLSERVER && requireIif(path, parameters);
+    return (sqlServerBool ? "iif(" : "")
+      + super.trans(target, esqlCon, path, parameters, env)
+      + (sqlServerBool ? ", 1, 0)" : "");
   }
 
   @Override